+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Allow an alchemy-field's field_title property to be set
 * Allow turning off specific column filters in alchemy-table
 * Add inline enum field view
+* Fix alchemy-pager icons
 
 ## 0.1.11 (2022-07-23)